[Performance] Add macros to adjust effects. If you have slow PC you can just switch off some details as light or day and night switching. There is also posibility to switch off collisions. It is necessary to handle with many vehicles on low machines. Instead of collisions is implemented FIFO in order to not allow vehicles move in the same time when green light is swiched.

+++ Ui/Checkpoints_manager/checkpoint-manager.cpp
@@ -29,7 +29,7 @@ Checkpoint* CheckpointManager::addCheckpoint( qreal x, qreal y, uint id, uint fl
 
     LOG_INFO( "Added new checkpoint with %i id", id );
 
-    connect( checkpoint, SIGNAL(checkpointReached(uint,uchar,Vehicle*)), this, SIGNAL(checkpointReached(uint,uchar,Vehicle*)) );
+    connect( checkpoint, SIGNAL(checkpointReached(uint,int,Vehicle*)), this, SIGNAL(checkpointReached(uint,int,Vehicle*)) );
 
     return checkpoint;
 }
